requirements.txt
axi~=0.1
click~=8.0.3
numpy~=1.22.0
shapely~=1.8.0
scipy~=1.7.3
tqdm~=4.62.3
pillow~=9.0.0

opensimplex~=0.3
requests~=2.25.1
matplotlib~=3.3.4